+++ packages/cosmwasm/types/index.d.ts
@@ -11,7 +11,6 @@ export {
   ContractDetails,
   CosmWasmClient,
   GetSequenceResult,
-  PostTxResult,
   SearchByHeightQuery,
   SearchByIdQuery,
   SearchBySentFromOrToQuery,